backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydub==0.25.1
moviepy==1.0.3
Pillow==10.1.0
scenedetect[opencv]==0.6.2
opencv-python>=4.8.0
ultralytics==8.1.0
google-generativeai
python-dotenv
numpy>=1.26.0,<2
torch<2.6.0
torchvision>=0.15.0
scikit-image>=0.21.0
transformers>=4.30.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
scipy>=1.11.0
